restructure folder layout to decouple code from page routes
