+++ __tests__/pages/index.js
@@ -1,13 +1,13 @@
 jest.mock('swr')
-jest.mock('../../pages/curriculum')
+jest.mock('../../containers/Curriculum')
 jest.mock('../../components/LandingPage')
 import React from 'react'
-import { render } from '@testing-library/react'
-import ReactDOM from 'react-dom'
-import Curriculum from '../../pages/curriculum'
-import IndexPage, {fetcher} from '../../pages/index'
-import LandingPage from '../../components/LandingPage'
 import useSWR from 'swr'
+import Curriculum from '../../containers/Curriculum'
+import Index from '../../containers/Index'
+import LandingPage from '../../components/LandingPage'
+import { render } from '@testing-library/react'
+import { fetcher } from '../../containers/Index'
 
 describe('Index Page', () => {
   test('fetcher should return a promise', async () => {
@@ -23,7 +23,7 @@ describe('Index Page', () => {
     })
     Curriculum.mockReturnValue( <h1>Hello Curriculum</h1> )
 
-    const { getByText } = render(<IndexPage />)
+    const { getByText } = render(<Index />)
     getByText('Hello Curriculum')
   })
   test('Should render landing page if user is not identified', async () => {
@@ -36,13 +36,13 @@ describe('Index Page', () => {
     })
     LandingPage.mockReturnValue(<h1>Hello Landing</h1>)
     
-    const { getByText } = render(<IndexPage />)
+    const { getByText } = render(<Index />)
     getByText('Hello Landing')
   })
   test('Should not render while page is loading', async () => {
     useSWR.mockReturnValue({})
     
-    const { container } = render(<IndexPage />)
+    const { container } = render(<Index />)
     expect(container).toMatchSnapshot()
   })
