+++ pages/_app.tsx
@@ -1,23 +1,3 @@
-// import libraries
-import * as React from 'react'
-import { ApolloProvider } from '@apollo/react-hooks'
-import { ApolloClient, NormalizedCacheObject } from 'apollo-boost'
-import { AppProps } from 'next/app'
-import withApollo from '../helpers/withApollo'
+import App from '../containers/App'
 
-// import scss
-import '../scss/index.scss'
-
-interface IProps extends AppProps {
-  apollo: ApolloClient<NormalizedCacheObject>
-}
-
-function MyApp({ Component, pageProps, apollo }: IProps) {
-  return (
-    <ApolloProvider client={apollo}>
-      <Component {...pageProps} />
-    </ApolloProvider>
-  )
-}
-
-export default withApollo(MyApp)
+export default App
