+++ pages/login.tsx
@@ -1,0 +1,3 @@
+import Login from '../containers/Login'
+
+export default Login
