+++ pages/signup.tsx
@@ -1,0 +1,3 @@
+import Signup from '../containers/Signup'
+
+export default Signup
